bme680(i): Change the altitude calculation slightly.
Based on the constants reccomended in
https://ncar.github.io/aircraft_ProcessingAlgorithms/www/PressureAltitude.pdf

Thanks to @bradcar for suggesting the change and providing the
reference. The changes to the result should be minor. To cite Brad:
"The more important thing is for people to set the nearest sea level
pressure (airport) at local time, or set the altitude to set the sea
level pressure."

Signed-off-by: robert-hh <robert@hammelrath.com>